+++ test/pom.xml
@@ -178,7 +178,7 @@ THE SOFTWARE.
     <dependency>
       <groupId>org.jenkins-ci.main</groupId>
       <artifactId>jenkins-test-harness</artifactId>
-      <version>2414.v185474555e66</version>
+      <version>2416.v986b_081cf07d</version>
       <scope>test</scope>
       <exclusions>
         <exclusion>
